+++ src/renderer/src/types/preload.d.ts
@@ -33,9 +33,13 @@ export interface IADBAPI {
     timeoutMs?: number,
   ) => Promise<AdbCommandResult>;
   runScrcpy: (deviceId: string, turnScreenOff: boolean) => Promise<any>;
+  openHardKeyboardSettings: (deviceId: string) => Promise<boolean>;
   connectWifi: (deviceId: string, ip: string) => Promise<any>;
   connectIp: (ip: string) => Promise<any>;
   pairDevice: (ipPort: string, code: string) => Promise<any>;
+  disconnectDevice: (
+    deviceId: string,
+  ) => Promise<{ success: boolean; message: string }>;
   getLocalIp: () => Promise<string>;
   getPackages: (
     deviceId: string,